resolve #2 1st step
Solution for the 1st step. The models used by the web-ui all have an
extra service 'modelinfo'. This service tells the web-ui what the base
model was. The map file should only contain the logical mappings, since
everything else is derived from the model as currently deployed on
Tengu. This allows a user to manually deploy extra services. These new
service will not be taken into account for the logical presentation of
the model, but they will be shown as regular services. Another plus, is
that the map file does not need to define shared services, since they
automatically will be shown. When a user removes a service that was part
of the logical mapping of the model, this node still will be shown in
the graph representation, but it's label will be set to
'NaS[{{service.name}}]'.